style: move test_moodleform into separate file
